+++ magda-migrator-registry-db/sql/V1_8__Add_webhook_retry_time_and_count.sql
@@ -1,0 +1,8 @@
+ALTER TABLE webhooks
+    ADD COLUMN enabled boolean NOT NULL DEFAULT true;
+
+ALTER TABLE webhooks
+    ADD COLUMN lastretrytime timestamptz NULL DEFAULT NULL;
+
+ALTER TABLE webhooks
+    ADD COLUMN retrycount integer NOT NULL DEFAULT 0;