+++ magda-registry-api/src/test/scala/au/csiro/data61/magda/registry/WebHookActorSpec.scala
@@ -1,12 +1,14 @@
 package au.csiro.data61.magda.registry
 
+import java.util.Timer
+
 import akka.actor.ActorRef
 import akka.http.scaladsl.model.StatusCodes
 import akka.pattern.ask
 import akka.util.Timeout
-import au.csiro.data61.magda.model.Registry.{ EventType, WebHook, WebHookConfig }
+import au.csiro.data61.magda.model.Registry.{EventType, WebHook, WebHookConfig}
 
-import scala.concurrent.Await
+import scala.concurrent.{Await, Promise}
 import scala.concurrent.duration._
 import akka.http.scaladsl.model.headers.RawHeader
 import com.auth0.jwt.JWT
@@ -55,6 +57,36 @@ class WebHookActorSpec extends ApiSpec {
     Await.result(actor ? WebHookActor.GetStatus("abc"), 5 seconds).asInstanceOf[WebHookActor.Status].isProcessing should not be (None)
   }
 
+  it("Will not creates a processor for newly-created disabled web hooks") { param =>
+    val actor = param.webHookActor
+    Await.result(actor ? WebHookActor.GetStatus("abc"), 5 seconds).asInstanceOf[WebHookActor.Status].isProcessing should be(None)
+
+    val hook = WebHook(
+      id = Some("abc"),
+      userId = None,
+      name = "abc",
+      active = true,
+      lastEvent = None,
+      url = "http://example.com/foo",
+      eventTypes = Set(EventType.CreateRecord),
+      isWaitingForResponse = None,
+      config = WebHookConfig(
+        optionalAspects = Some(List("aspect")),
+        includeEvents = Some(true),
+        includeRecords = Some(true),
+        includeAspectDefinitions = Some(true),
+        dereference = Some(true)),
+      enabled = false)
+
+    param.asAdmin(Post("/v0/hooks", hook)) ~> param.api.routes ~> check {
+      status shouldEqual StatusCodes.OK
+    }
+
+    processAndWaitUntilDone(actor)
+
+    Await.result(actor ? WebHookActor.GetStatus("abc"), 5 seconds).asInstanceOf[WebHookActor.Status].isProcessing should be (None)
+  }
+
   it("removes the processor for removed web hooks") { param =>
     val actor = param.webHookActor
     val hook = WebHook(
@@ -91,4 +123,5 @@ class WebHookActorSpec extends ApiSpec {
 
     Await.result(actor ? WebHookActor.GetStatus("abc"), 5 seconds).asInstanceOf[WebHookActor.Status].isProcessing should be(None)
   }
+
 }